fix: Add props to alias query
There is an edge case where the page that aliases point to was created
_after_ the page with the name of the alias, giving the alias pointer page
a lower number and therefore selected first in queries.

An easy way (but not foolproof) way to reproduce this problem would be to
create a page in the plural (e.g. `[[people]]`), then decide to make
a singular page as the main page (e.g. `[[person]]` with
`alias:: people`). Any other sort of renaming or reassigning could lead to
this edge case, too.

Previously, this query worked by assuming that the first page created would
have all other pages be aliased to it later.

This fix works by assuming that _only_ the page that all other aliases
point to contains any properties. By querying for the mere presence of
properties, we can deduce which page all the others are aliases of.